update on js-sdk with examples
