+++ examples/create-react-app/src/components/Header/Header.js
@@ -1,5 +1,5 @@
 import { withGrandeur } from "grandeur-js/react";
-import { Link, useNavigate } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 function Header(props) {
   const navigate = useNavigate();
@@ -15,15 +15,13 @@ function Header(props) {
       <div className="text-gray-600 font-sans text-md font-bold">
         Internet Switch
       </div>
-      <Link to={"/"}>
-        <button
-          id="logout"
-          onClick={logout}
-          className="h-8 text-gray-600 bg-gray-200 rounded-md px-5 text-sm font-sans font-bold cursor-pointer"
-        >
-          Logout
-        </button>
-      </Link>
+      <button
+        id="logout"
+        onClick={logout}
+        className="h-8 text-gray-600 bg-gray-200 rounded-md px-5 text-sm font-sans font-bold cursor-pointer"
+      >
+        Logout
+      </button>
     </div>
   );
 }
